requirements.txt
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.34.0
pandas==2.3.3
numpy==2.3.5
pyarrow==22.0.0
db-dtypes==1.4.4
pandas-gbq==0.31.0
lifelines==0.30.0
statsmodels==0.14.5
scipy==1.16.3
scikit-learn==1.7.2
matplotlib==3.10.7
seaborn==0.13.2
tabulate==0.9.0
